part1/flask-frontend/requirements.txt
Flask-WTF
Flask-Bootstrap4
requests
